Add files via upload
This project is part of my second-week task during an ongoing internship with FlexiSAF Edusoft, focused on Generative AI and Data Science. It involves analyzing and visualizing company sales data using Python, Pandas, and Matplotlib. The dataset contains monthly sales records of various products. The project includes a line plot to visualize the total monthly profit, as well as subplots to compare the monthly sales of bathing soap and facewash. The visualizations are clean, well-labeled, and easy to interpret, showcasing practical data storytelling skills. This project demonstrates my growing proficiency in data analysis and Python-based visualization.